[IMP] sale_margin_percentage: separate one model per file
To follow our good practices each model should be in its own file with
the same name as the model.
