+++ sale_margin_percentage/models/__init__.py
@@ -1,3 +1,4 @@
 from . import res_company
 from . import res_config_settings
 from . import sale_order
+from . import sale_order_line
